Rename edge to edges
